+++ CHANGES.md
@@ -1,5 +1,8 @@
+Updated ROCm to 7.2.3
 Added Kimodo
 Added hipfire
+Added llama-cpp-turboquant
+Added Z-Anime to ComfyUI
 Updated llama.cpp
 Updated SillyTavern
 Updated WhisperSpeech web UI